# api.yml
swagger: '2.0'
info:
  title: Shop Example API
  version: "0.1"
  description: Simple example API
consumes:
  - application/json
produces:
  - application/json
basePath: "/api"
securityDefinitions:
  jwt:
    type: apiKey
    name: Authorization
    in: header
    x-authentication-scheme: Bearer
    x-bearerInfoFunc: app.decode_token
paths:
  /test1/{test1_id}:
    get:
      security:
        - jwt: [ 'secret' ]
      tags: [ Test1 ]
      operationId: app.get_test1
      summary: Testing endpoint
      parameters:
        - $ref: '#/parameters/test1_id'
      responses:
        200:
          description: Return test1
        404:
          description: Does not exist
  /person/:
    post:
      tags: [ Person ]
      operationId: app.person_add
      parameters:
        - $ref: '#/parameters/person_body'
      responses:
        200:
          description: Return status
        409:
          description: Already exists
  /person/{person_name}:
    get:
      tags: [ Person ]
      operationId: app.person_find
      parameters:
        - $ref: '#/parameters/person_name'
      responses:
        200:
          description: Return person object
        404:
          description: Person not found
  /auth:
    post:
      tags: [ Auth ]
      operationId: app.auth
      parameters:
        - $ref: '#/parameters/auth_body'
      responses:
        200:
          description: Return jwt token
        401:
          description: Bad auth

  /auth_microservice:
    post:
      tags: [ Auth ]
      operationId: app.auth_microservice
      parameters:
        - $ref: '#/parameters/auth_body_microservice'
      responses:
        200:
          description: Return jwt token
        401:
          description: Bad auth

parameters:
  test1_id:
    name: test1_id
    description: Test1 unique id
    in: path
    type: string
    required: true
    pattern: "^[a-zA-Z0-9-]+$"

  person_body:
    name: person_body
    description: Test1 unique id
    in: body
    schema:
      required:
        - name
      properties:
        name:
          type: string
        surname:
          type: string

  auth_body:
    name: auth_body
    description: Login and get token
    in: body
    schema:
      required:
        - username
        - password
      properties:
        username:
          type: string
        password:
          type: string
  auth_body_microservice:
    name: auth_body_microservice
    description: Login and get token for microservices
    in: body
    schema:
      required:
        - apikey
      properties:
        apikey:
          type: string

  person_name:
    name: person_name
    description: The name of the person we're looking for
    in: path
    required: true
    type: string